Recover malformed JSON instead of text-parsing it
A structured reply that failed JSON.parse fell through to
parseReceiptText, which reads "NAME  12.34" rows. Pretty-printed JSON
matches that shape, so `  "price": -3.00` became an item literally named
`"price":`, and the indentation nested it as a child. Only the last
property in each object matched — the rest carry a trailing comma — so a
whole receipt collapsed into a handful of plausible-looking amounts.

Handle the two ways a reply misses bare JSON: strip a markdown fence or
prose preamble by scanning from the first brace, and close a reply
truncated by max_tokens back to the last value that completed. Losing a
40-line receipt over its last row helps nobody.

Gate the text fallback so it never sees a reply meant to be structured,
returning no items instead, which routes the client to local OCR. Raise
max_tokens to 4096, since a cut lands past the items on tax and tip.

Move the schema and its parsing into utils/receiptSchema.ts so they are
testable outside the route.
